requirements.txt
Appium-Python-Client==4.1.0
pytest==8.3.1
allure-pytest==2.13.5